Split ext.sprite into smaller modules
