+++ sdl2/ext/__init__.py
@@ -18,6 +18,8 @@ from .gui import *
 from .msgbox import *
 from .image import *
 from .pixelaccess import *
+from .renderer import *
 from .sprite import *
+from .spritesystem import *
 from .surface import *
 from .window import *
